# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
        "/api/admins": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List of Admins",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "27f6447f654b54ee491274c73668fa10",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/admins/register": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create Admin",
                "operationId": "030392af54b96e99a21d62b9cce7f86f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "Ahrrabi@example.com",
                                    "username": "admin",
                                    "password": "Ahrrabi1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "admin@example.com"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "$2y$10$Gs3IpiZd4KkmyVF8tSI55ujEoAKe0.eXqQK.QklsERNMHbtfqR2uq"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admins/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Login",
                "operationId": "a6f00a3382b138008f0d15030d5375fd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "description": "Email address of the admin",
                                                "type": "string"
                                            },
                                            "password": {
                                                "description": "Password of the admin",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "Ahrrabi@example.com",
                                    "password": "Ahrrabi1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "email": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The email field is required."
                                            }
                                        },
                                        "password": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The password field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admins/update/{admin}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update Admin",
                "operationId": "73d359361bb911cb9070cc0fde357bcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "abdou@gmail.com",
                                    "username": "adbo.ah",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 5
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "abdou@gmail.com"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "adbo.ah"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-12-28T21:23:09.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-12-28T21:34:47.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admins/delete/{admin}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Admin",
                "operationId": "8d4028743fb30861d1bc542c0883800f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete admin success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}